fix: remove unnecessary cast in turso_client cleanup function
Remove unnecessary u64 -> u64 cast in cleanup_old_data function.
The execute() method already returns u64, so the cast is redundant.

+++ src/storage/turso_client.rs
@@ -516,7 +516,7 @@ impl TursoClient {
 
         let rows_affected = result;
         info!("Cleaned up {} old weather data records", rows_affected);
-        Ok(rows_affected as u64)
+        Ok(rows_affected)
     }
 }
